fix(web): add location modal invisible on safari
